server/app/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
requests
django-environ